clean up logging tests
